+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Allow creating `DOMTokenList` instances on the fly with `Hawkejs.DOMTokenList`
 * Also call `attributeChangedCallback` after removing an attribute on the server side
 * Add some missing methods & properties to the server-side `DOMTokenList` implementation
+* Add support for custom `DOMTokenList` attributes
 
 ## 2.2.17 (2022-10-22)